# Conflicts:
#	core/src/main/java/com/orientechnologies/orient/core/db/document/ODatabaseDocumentTx.java
#	core/src/test/java/com/orientechnologies/orient/core/sql/OCommandExecutorSQLSelectTest.java

+++ distributed/src/test/java/com/orientechnologies/orient/server/distributed/scenariotest/DeleteAndLazarusScenarioTest.java
@@ -37,16 +37,16 @@ import static org.junit.Assert.*;
  * - server3 joins the cluster (restart)
  * - shutdown server1 (so quorum for CRUD operation on r1 will not be reached)
  * - delete request for r1 on server3:
- *      - quorum not reached because r1* on server3 is not consistent with r1 on server2 (different values and versions)
- *      - delete operation is aborted on server2 and is rolled back on server3 (resurrection)
+ * - quorum not reached because r1* on server3 is not consistent with r1 on server2 (different values and versions)
+ * - delete operation is aborted on server2 and is rolled back on server3 (resurrection)
  * - restart server1 (so quorum for CRUD operation on r1 will be reached)
  * - check consistency: r1 is still present on server1 and server2, and r1* is present on server3.
  * - delete request for r1 on server1:
- *      - quorum reached
- *      - check consistency: r1 is not present on server1 and server2, and r1* is not present on server3.
+ * - quorum reached
+ * - check consistency: r1 is not present on server1 and server2, and r1* is not present on server3.
  *
  * @author Gabriele Ponzi
- * @email  <gabriele.ponzi--at--gmail.com>
+ * @email <gabriele.ponzi--at--gmail.com>
  */
 
 public class DeleteAndLazarusScenarioTest extends AbstractScenarioTest {
@@ -134,9 +134,9 @@ public class DeleteAndLazarusScenarioTest extends AbstractScenarioTest {
     ODatabaseDocumentTx dbServer3 = null;
     try {
       r1onServer3 = retrieveRecord(getPlocalDatabaseURL(serverInstance.get(2)), "R001");
-      dbServer3 = new ODatabaseDocumentTx(getPlocalDatabaseURL(serverInstance.get(2))).open("admin","admin");
-      r1onServer3.field("firstName","Darth");
-      r1onServer3.field("lastName","Vader");
+      dbServer3 = new ODatabaseDocumentTx(getPlocalDatabaseURL(serverInstance.get(2))).open("admin", "admin");
+      r1onServer3.field("firstName", "Darth");
+      r1onServer3.field("lastName", "Vader");
       r1onServer3.save();
       System.out.println(r1onServer3.getRecord().toString());
     } catch (Exception e) {
@@ -161,7 +161,7 @@ public class DeleteAndLazarusScenarioTest extends AbstractScenarioTest {
     r1onServer1 = retrieveRecord(getDatabaseURL(serverInstance.get(0)), "R001");
     r1onServer2 = retrieveRecord(getDatabaseURL(serverInstance.get(1)), "R001");
 
-    assertEquals(1, r1onServer1.field("@version"));
+    assertEquals((Integer) 1, r1onServer1.field("@version"));
     assertEquals("R001", r1onServer1.field("id"));
     assertEquals("Luke", r1onServer1.field("firstName"));
     assertEquals("Skywalker", r1onServer1.field("lastName"));
@@ -175,7 +175,7 @@ public class DeleteAndLazarusScenarioTest extends AbstractScenarioTest {
     assertEquals("R001", r1onServer3.field("id"));
     assertEquals("Darth", r1onServer3.field("firstName"));
     assertEquals("Vader", r1onServer3.field("lastName"));
-    assertEquals(initialVersion+1, r1onServer3.field("@version"));
+    assertEquals((Integer) (initialVersion + 1), r1onServer3.field("@version"));
 
     // shutdown server1
     System.out.println("Network fault on server1.\n");
@@ -186,7 +186,7 @@ public class DeleteAndLazarusScenarioTest extends AbstractScenarioTest {
     try {
       dbServer3 = poolFactory.get(getDatabaseURL(serverInstance.get(2)), "admin", "admin").acquire();
       dbServer3.command(new OCommandSQL("delete from Person where @rid=#27:0")).execute();
-    } catch(Exception e) {
+    } catch (Exception e) {
       System.out.println(e.getMessage());
     }
 
@@ -199,7 +199,7 @@ public class DeleteAndLazarusScenarioTest extends AbstractScenarioTest {
     r1onServer1 = retrieveRecord(getDatabaseURL(serverInstance.get(0)), "R001");
     r1onServer2 = retrieveRecord(getDatabaseURL(serverInstance.get(1)), "R001");
 
-    assertEquals(1, r1onServer1.field("@version"));
+    assertEquals((Integer) 1, r1onServer1.field("@version"));
     assertEquals("R001", r1onServer1.field("id"));
     assertEquals("Luke", r1onServer1.field("firstName"));
     assertEquals("Skywalker", r1onServer1.field("lastName"));
@@ -212,7 +212,7 @@ public class DeleteAndLazarusScenarioTest extends AbstractScenarioTest {
     // r1* is still present on server3
     r1onServer3 = retrieveRecord(getDatabaseURL(serverInstance.get(2)), "R001");
 
-    assertEquals(2, r1onServer3.field("@version"));
+    assertEquals((Integer) 2, r1onServer3.field("@version"));
     assertEquals("R001", r1onServer3.field("id"));
     assertEquals("Darth", r1onServer3.field("firstName"));
     assertEquals("Vader", r1onServer3.field("lastName"));
@@ -221,7 +221,7 @@ public class DeleteAndLazarusScenarioTest extends AbstractScenarioTest {
     try {
       dbServer1 = poolFactory.get(getRemoteDatabaseURL(serverInstance.get(0)), "admin", "admin").acquire();
       Integer result = dbServer1.command(new OCommandSQL("delete from " + r1Rid)).execute();
-    } catch(Exception e) {
+    } catch (Exception e) {
       e.printStackTrace();
     }
